+++ internal/test/common/src/main/kotlin/com/github/jcornaz/miop/internal/test/ManualTimer.kt
@@ -1,0 +1,9 @@
+package com.github.jcornaz.miop.internal.test
+
+expect class ManualTimer {
+    suspend fun await(time: Int)
+
+    fun advanceTo(time: Int = 1)
+
+    suspend fun terminate()
+}